moving fog classes to separete namespace and folder
Fixing vcloud-tools folder structure and namespace

moved fog related classes under Vcloud::Fog namespace and inside
lib/vcloud/fog directory
