--- lib/vcloud/core/content_types.rb
@@ -1,9 +1,0 @@
-module Vcloud
-  module ContentTypes
-    CATALOG = 'application/vnd.vmware.vcloud.catalog+xml'
-    ORG = 'application/vnd.vmware.vcloud.org+xml'
-    VDC = 'application/vnd.vmware.vcloud.vdc+xml'
-    CATALOG_ITEM = 'application/vnd.vmware.vcloud.catalogItem+xml'
-    NETWORK = 'application/vnd.vmware.vcloud.network+xml'
-  end
-end
