--- lib/vcloud/core/relation.rb
@@ -1,6 +1,0 @@
-module Vcloud
-  module RELATION
-    PARENT = 'up'
-    CHILD = 'down'
-  end
-end
