+++ lib/vcloud/core/vapp_template.rb
@@ -7,7 +7,7 @@ module Vcloud
 
     def self.get catalog_name, catalog_item_name
       raise "provide catalog and catalog item name to load vappTemplate" unless catalog_name && catalog_item_name
-      @vcloud_attributes = FogServiceInterface.new.template(catalog_name, catalog_item_name)
+      @vcloud_attributes = Fog::ServiceInterface.new.template(catalog_name, catalog_item_name)
       raise 'Could not find template vApp' unless @vcloud_attributes
       new(@vcloud_attributes)
     end
