+++ lib/vcloud/fog.rb
@@ -1,0 +1,5 @@
+require 'fog'
+require 'vcloud/fog/content_types'
+require 'vcloud/fog/relation'
+require 'vcloud/fog/service_interface'
+require 'vcloud/fog/model_interface'
