+++ lib/vcloud/fog/content_types.rb
@@ -1,0 +1,11 @@
+module Vcloud
+  module Fog
+    module ContentTypes
+      CATALOG = 'application/vnd.vmware.vcloud.catalog+xml'
+      ORG = 'application/vnd.vmware.vcloud.org+xml'
+      VDC = 'application/vnd.vmware.vcloud.vdc+xml'
+      CATALOG_ITEM = 'application/vnd.vmware.vcloud.catalogItem+xml'
+      NETWORK = 'application/vnd.vmware.vcloud.network+xml'
+    end
+  end
+end
