+++ lib/vcloud/fog/relation.rb
@@ -1,0 +1,8 @@
+module Vcloud
+  module Fog
+    module RELATION
+      PARENT = 'up'
+      CHILD = 'down'
+    end
+  end
+end
